1. Modify the signUp handler method: Update the signUp handler to include a User object so that user information can be populated into the form fields.
2. Create a registerUser POST handler method:
Implement a new registerUser method to handle form submissions from the
signup page. This method should:

Extract the user information from the form.

Save the user data into the database.

Use a try-catch block to handle any exceptions that occur during
registration.

Store error or success messages in the HttpSession to display them as
alerts on the signup page.

3. Create a Message class in the helper package:
Add a Message class to the helper package. This class will be used to
create structured alert messages (such as success, error, or warning) to
be displayed on the web page.